tests: add LULESH to weekly regression
LULESH is a popular GPU HPC application that acts as a good test
for several memory and compute patterns.  Thus, including it in
the weekly regressions will help verify correctness and
functionality for code that affects the GPU.  The default LULESH
input runs 10 iterations and takes 3-4 hours.  Hence, it is not
appropriate for nightly regressions.

Change-Id: Ic1b73ab32fdd5cb1b973f2676b272adb91b2a98e
Reviewed-on: https://gem5-review.googlesource.com/c/public/gem5/+/50952
Maintainer: Matt Sinclair <mattdsinclair@gmail.com>
Tested-by: kokoro <noreply+kokoro@google.com>
Reviewed-by: Bobby R. Bruce <bbruce@ucdavis.edu>